requirements.txt
openai
python-dotenv
gunicorn
psycopg2-binary
requests
django-social-share